[FIX] Missing functions on_attach_callback and on_detach_callback in report_backend widget.
These missing two function in the widget cause two bugs in the Odoo
Client. (tested on Odoo Enterprise Version).

1. The bug occor when we are on VAT Report view page then "click" on "Home
menu" (located on the top left of the screen). This bug is caused by
missing of on_detach_callback.

2. The bug occor when we are on VAT Report view page and then we
"refresh" the page (pressing F5 or CTRL+R). This bug is caused by
missing of on_attach_callback.